requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
joblib>=1.0.0
pillow>=8.2.0